config was getting large - refactor
