+++ src/config/mod.rs
@@ -1,7 +1,6 @@
-//! config.rs - load/parse a ZPL configuration TOML file
+//! Loads and parses a ZPL TOML configuration file into a `Config` struct.
 
 use std::collections::HashMap;
-use std::collections::HashSet;
 use std::net::IpAddr;
 use std::path::Path;
 use std::path::PathBuf;
@@ -13,11 +12,19 @@ use toml::Table;
 use crate::context::CompilationCtx;
 use crate::crypto::sha256;
 use crate::errors::CompilationError;
-use crate::protocols::{
-    IanaProtocol, IcmpFlowType, PortSpec, Protocol, ProtocolError, ZPR_L7_BUILTINS,
-};
+use crate::protocols::{IcmpFlowType, PortSpec, Protocol, ProtocolError, ZPR_L7_BUILTINS};
 use crate::zpl;
 use zpr::policy_types::Attribute;
+
+mod node_link;
+mod protocol;
+mod service;
+mod trusted_service;
+
+use node_link::{parse_link, parse_node, parse_substrate_addrs};
+use protocol::parse_protocol;
+use service::parse_service;
+use trusted_service::parse_trusted_service;
 
 /// Helper to create a ConfigError. Works with a single string (or &str) argument
 /// (really anything that has a to_string function), or with two args: a format string and arguments.
@@ -109,10 +116,10 @@ pub struct Link {
     pub peer_a_id: String,
 
     // Interfaces map to Node substrate interface IDs.
-    pub peer_a_interface: Option<String>,
+    pub peer_a_interface: String,
 
     pub peer_b_id: String,
-    pub peer_b_interface: Option<String>,
+    pub peer_b_interface: String,
 
     // Link always gets 'zpr.cost' attribute at least.
     pub attributes: Vec<(String, String)>,
@@ -568,217 +575,6 @@ impl ConfigParse {
     }
 }
 
-fn require_key(ctx: &str, table: &Table, key: &str) -> Result<(), CompilationError> {
-    if !table.contains_key(key) {
-        return Err(err_config!("error in {}: missing entry for {}", ctx, key));
-    }
-    Ok(())
-}
-
-/// Parse a single node table.
-fn parse_node(node_id: &str, node: &Table, ctx: &CompilationCtx) -> Result<Node, CompilationError> {
-    warn_unknown_node_property(node, ctx)?;
-    require_key(&format!("nodes.{}", node_id), node, "zpr_address")?;
-    let zpr_address = node["zpr_address"]
-        .as_str()
-        .ok_or(err_config!("node {} invalid zpr_address", node_id))?
-        .to_string();
-
-    let provider = parse_provider(&format!("node {}", node_id), node)?;
-
-    Ok(Node {
-        id: node_id.to_string(),
-        zpr_address,
-        substrate_addrs: None, // filled in later if present
-        provider,
-    })
-}
-
-/// Parses the "link.<LINKID>" table from the toml config.
-/// This should include a peers entry (2 element array) and an optional attributes entry.
-pub fn parse_link(
-    link_id: &str,
-    link_tbl: &Table,
-    ctx: &CompilationCtx,
-    nodes: &HashMap<String, Node>,
-) -> Result<Link, CompilationError> {
-    warn_unknown_link_property(link_tbl, ctx)?;
-
-    require_key(&format!("links.{}", link_id), link_tbl, "peers")?;
-    let peers = link_tbl["peers"]
-        .as_array()
-        .ok_or(err_config!("link {} peers should be an array", link_id))?;
-
-    if peers.len() != 2 {
-        return Err(err_config!("link {} must have exactly two peers", link_id));
-    }
-
-    let mut peer_a_id = None;
-    let mut peer_a_interface = None;
-    let mut peer_b_id = None;
-    let mut peer_b_interface = None;
-
-    for (i, peer) in peers.iter().enumerate() {
-        let peer_tbl =
-            peer.as_table()
-                .ok_or(err_config!("link {} peer {} should be a table", link_id, i))?;
-
-        require_key(&format!("links.{}.peers[{}]", link_id, i), peer_tbl, "node")?;
-        let node_id = peer_tbl["node"]
-            .as_str()
-            .ok_or(err_config!(
-                "link {} peer {} node should be a string",
-                link_id,
-                i
-            ))?
-            .to_string();
-
-        let refnode = nodes.get(&node_id);
-        if !refnode.is_some() {
-            return Err(err_config!(
-                "link {} peer {} references unknown node '{}'",
-                link_id,
-                i,
-                node_id
-            ));
-        }
-        let refnode = refnode.unwrap();
-
-        if refnode.substrate_addrs.is_none() {
-            return Err(err_config!(
-                "link {} peer {} references node '{}' with no substrate_addrs",
-                link_id,
-                i,
-                node_id
-            ));
-        }
-
-        let interface = if peer_tbl.contains_key("interface") {
-            let iname = peer_tbl["interface"]
-                .as_str()
-                .ok_or(err_config!(
-                    "link {} peer {} interface should be a string",
-                    link_id,
-                    i
-                ))?
-                .to_string();
-            // Interface name must exist on node.
-            if !refnode
-                .substrate_addrs
-                .as_ref()
-                .unwrap()
-                .contains_key(&iname)
-            {
-                return Err(err_config!(
-                    "link {} peer {} references interface '{}' which does not exist on node '{}'",
-                    link_id,
-                    i,
-                    iname,
-                    node_id
-                ));
-            }
-            Some(iname)
-        } else {
-            // Interface can be omitted if node has single substreate address.
-            if refnode.substrate_addrs.as_ref().unwrap().len() == 1 {
-                // Just take the single interface.
-                refnode
-                    .substrate_addrs
-                    .as_ref()
-                    .unwrap()
-                    .keys()
-                    .next()
-                    .cloned()
-            } else {
-                return Err(err_config!(
-                    "link {} peer {} missing interface and node '{}' has multiple substrate_addrs",
-                    link_id,
-                    i,
-                    node_id
-                ));
-            }
-        };
-
-        if i == 0 {
-            peer_a_id = Some(node_id);
-            peer_a_interface = interface;
-        } else {
-            peer_b_id = Some(node_id);
-            peer_b_interface = interface;
-        }
-    }
-
-    let attrs = match parse_attribute_tuples(&format!("links.{}", link_id), link_tbl, "attributes")?
-    {
-        Some(attrs) => attrs,
-        None => Vec::new(),
-    };
-
-    Ok(Link {
-        peer_a_id: peer_a_id.unwrap(),
-        peer_a_interface,
-        peer_b_id: peer_b_id.unwrap(),
-        peer_b_interface,
-        attributes: attrs,
-    })
-}
-
-/// Parse a node substrate addrs table.
-fn parse_substrate_addrs(
-    node_id: &str,
-    addrs: &Table,
-    _ctx: &CompilationCtx,
-) -> Result<HashMap<String, Interface>, CompilationError> {
-    let mut substrate_addrs = HashMap::new();
-    for (ifname, v) in addrs {
-        let iface = parse_interface(
-            ifname,
-            v.as_str().ok_or(err_config!(
-                "node {} substrate_addr {} should be a \"HOST:PORT\" formatted string",
-                node_id,
-                ifname
-            ))?,
-        )?;
-        substrate_addrs.insert(ifname.to_string(), iface);
-    }
-    Ok(substrate_addrs)
-}
-
-fn warn_unknown_node_property(node: &Table, ctx: &CompilationCtx) -> Result<(), CompilationError> {
-    for elem in node.keys() {
-        match elem.as_str() {
-            "provider" => (),
-            "zpr_address" => (),
-            "substrate_addrs" => (),
-            _ => ctx.warn(&format!(
-                "unknown property '{elem}' detected while parsing node",
-            ))?,
-        }
-    }
-
-    Ok(())
-}
-
-fn warn_unknown_link_property(link: &Table, ctx: &CompilationCtx) -> Result<(), CompilationError> {
-    for elem in link.keys() {
-        match elem.as_str() {
-            "peers" => (),
-            "attributes" => (),
-            _ => ctx.warn(&format!(
-                "unknown property '{elem}' detected while parsing link",
-            ))?,
-        }
-    }
-    Ok(())
-}
-
-fn parse_provider(ctx: &str, table: &Table) -> Result<Vec<(String, String)>, CompilationError> {
-    match parse_attribute_tuples(ctx, table, "provider")? {
-        Some(attrs) => Ok(attrs),
-        None => Err(err_config!("{} missing provider", ctx)),
-    }
-}
-
 /// Parse attribute tuples from a TOML table field.
 ///
 /// This is how we generally encode ZPR attriutes into the toml, for example:
@@ -793,7 +589,7 @@ fn parse_provider(ctx: &str, table: &Table) -> Result<Vec<(String, String)>, Com
 /// and this will return a vector of (key, value) tuples.
 ///
 /// If key is not found, returns None.
-fn parse_attribute_tuples(
+pub(super) fn parse_attribute_tuples(
     ctx: &str,
     table: &Table,
     field_name: &str,
@@ -809,39 +605,13 @@ fn parse_attribute_tuples(
     Ok(Some(attrs))
 }
 
-/// Parse the nodes interface entry which is just "HOST:PORT"
-fn parse_interface(ifname: &str, netaddr: &str) -> Result<Interface, CompilationError> {
-    // Form of `netaddr` is HOST:PORT, host may be a hostname (which may need to be run through the resolver)
-    // or an IPv4 or IPv6 address.
-
-    // We'll try to parse as a SocketAddr first (which requires an IP address, not a name)
-    //let saddr: std::net::SocketAddr = netaddr.parse();
-    match netaddr.parse::<std::net::SocketAddr>() {
-        Ok(saddr) => Ok(Interface {
-            host: saddr.ip().to_string(),
-            port: saddr.port(),
-        }),
-        Err(_) => {
-            // Did not parse as a SocketAddr, so try as "hostname:portnum"
-            let parts: Vec<&str> = netaddr.split(':').collect();
-            if parts.len() != 2 {
-                return Err(err_config!(
-                    "interface {} netaddr must be in the form HOST:PORT",
-                    ifname
-                ));
-            }
-            let portnum = parts[1].parse::<u16>().map_err(|_| {
-                err_config!(
-                    "interface {} port number is not a valid: {}",
-                    ifname,
-                    parts[1]
-                )
-            })?;
-            Ok(Interface {
-                host: parts[0].to_string(),
-                port: portnum,
-            })
-        }
+pub(super) fn parse_provider(
+    ctx: &str,
+    table: &Table,
+) -> Result<Vec<(String, String)>, CompilationError> {
+    match parse_attribute_tuples(ctx, table, "provider")? {
+        Some(attrs) => Ok(attrs),
+        None => Err(err_config!("{} missing provider", ctx)),
     }
 }
 
@@ -874,476 +644,10 @@ fn tuples_to_tuple_str_vec(
     Ok(svec)
 }
 
-// Parse an individual trusted_service table.
-fn parse_trusted_service(
-    ts_id: &str,
-    ts: &Table,
-    ctx: &CompilationCtx,
-) -> Result<TrustedService, CompilationError> {
-    warn_unknown_ts_property(ts, ctx)?;
-    // The "api" value is optional for the default trusted service.
-    let mut is_default = false;
-    let api = if ts.contains_key("api") {
-        ts["api"]
-            .as_str()
-            .ok_or(err_config!("trusted_service {} missing api", ts_id))?
-            .to_string()
-    } else if ts_id == zpl::DEFAULT_TRUSTED_SERVICE_ID {
-        is_default = true;
-        zpl::DEFAULT_TRUSTED_SERVICE_API.to_string()
-    } else {
-        return Err(err_config!("trusted_service {} missing api", ts_id));
-    };
-    let cert_path = if ts.contains_key("cert_path") {
-        Some(PathBuf::from(ts["cert_path"].as_str().ok_or(
-            err_config!("trusted_service {} cert_path is not a string", ts_id),
-        )?))
-    } else if is_default {
-        // The path is the only thing required for the default section.
-        ctx.warn("no cert_path for default trusted_service, certificate checking disabled")?;
-        None
-    } else {
-        None
-    };
-
-    let returns_attrs: Vec<String>;
-    let identity_attrs: Vec<String>;
-    let client_svc: Option<String>;
-    let service_svc: Option<String>;
-    if !is_default {
-        returns_attrs = parse_string_array(ts, "returns_attributes", "trusted_service")?;
-        identity_attrs = parse_string_array(ts, "identity_attributes", "trusted_service")?;
-
-        if ts.contains_key("client") {
-            client_svc = Some(
-                ts["client"]
-                    .as_str()
-                    .ok_or(err_config!("trusted_service {} client parse error", ts_id))?
-                    .to_string(),
-            );
-        } else {
-            client_svc = Some(format!("{}-client", ts_id));
-        }
-        if ts.contains_key("service") {
-            service_svc = Some(
-                ts["service"]
-                    .as_str()
-                    .ok_or(err_config!("trusted_service {} service parse error", ts_id))?
-                    .to_string(),
-            );
-        } else {
-            service_svc = Some(format!("{}-vs", ts_id));
-        }
-    } else {
-        if ts.contains_key("returns_attributes") {
-            return Err(err_config!(
-                "default trusted_service does not allow custom returns_attributes"
-            ));
-        }
-        if ts.contains_key("identity_attributes") {
-            return Err(err_config!(
-                "default trusted_service does not allow custom identity_attributes"
-            ));
-        }
-        returns_attrs = vec![format!("{} -> {}", zpl::KATTR_CN, zpl::KATTR_CN)];
-        identity_attrs = vec![String::from(zpl::KATTR_CN)];
-        client_svc = None;
-        service_svc = None;
-    }
-
-    let mut returns = HashMap::new();
-    for ra in &returns_attrs {
-        let (service_key_name, zpr_attr) = parse_attribute_mapping(ra)?;
-        if returns.contains_key(&service_key_name) {
-            return Err(err_config!(
-                "trusted_service {} contains duplicate service attribute name '{}'",
-                ts_id,
-                service_key_name
-            ));
-        }
-        returns.insert(service_key_name, zpr_attr);
-    }
-
-    let mut idents = Vec::new();
-    for ra in &identity_attrs {
-        // The ident attribute (for now) must exist in the returns attributes.
-        if !returns.contains_key(ra) {
-            return Err(err_config!(
-                "trusted_service {} identity attribute '{}' not in returns_attributes",
-                ts_id,
-                ra
-            ));
-        }
-        idents.push(ra.to_string());
-    }
-
-    let provider = if ts.contains_key("provider") {
-        Some(parse_provider(&format!("trusted_service {ts_id}"), ts)?)
-    } else {
-        if !is_default {
-            return Err(err_config!("trusted_service {} missing provider", ts_id));
-        }
-        None
-    };
-
-    Ok(TrustedService {
-        id: ts_id.to_string(),
-        api,
-        cert_path,
-        returns_attrs: returns,
-        identity_attrs: idents,
-        provider,
-        client: client_svc,
-        service: service_svc,
-    })
-}
-
-/// The mapping string format is "<service-key-name> -> <attribute-spec>" where attribute
-/// spec is:
-///   - <class-name>.<attribute-name> for a regular single value attribute.
-///   - #<class-name>.<attribute-name> for a tag attribute.
-///   - <class-name>.<attribute-name>{} for a multi-valued attribute.
-///
-/// Note that we never use "optional" flag in ZPLC.
-fn parse_attribute_mapping(mapping: &str) -> Result<(String, Attribute), CompilationError> {
-    let parts: Vec<&str> = mapping.split("->").collect();
-    if parts.len() != 2 {
-        return Err(err_config!(
-            "invalid attribute mapping '{}', must be of the form '<service-key-name> -> <attribute-spec>'",
-            mapping
-        ));
-    }
-    let service_key_name = parts[0].trim().to_string();
-    let attr_spec = parts[1].trim();
-
-    let zpr_attr = if let Some(stripped) = attr_spec.strip_prefix("#") {
-        Attribute::tag(stripped).build()?
-    } else if let Some(stripped) = attr_spec.strip_suffix("{}") {
-        Attribute::tuple(stripped).multi().build()?
-    } else {
-        Attribute::tuple(attr_spec).single().build()?
-    };
-
-    // In theory we can support any attribute names if they are quoted.
-    // But until we support that on VS we will not permit some characters
-    // here.
-    let valid_chars: HashSet<char> =
-        "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-./"
-            .chars()
-            .collect();
-    for c in zpr_attr.zpl_key().chars() {
-        if !valid_chars.contains(&c) {
-            return Err(err_config!(
-                "invalid attribute name '{}' in mapping '{}', contains invalid character '{}'",
-                zpr_attr.zpl_key(),
-                mapping,
-                c
-            ));
-        }
-    }
-
-    Ok((service_key_name, zpr_attr))
-}
-
-fn warn_unknown_ts_property(ts: &Table, ctx: &CompilationCtx) -> Result<(), CompilationError> {
-    for elem in ts.keys() {
-        match elem.as_str() {
-            "cert_path" => (),
-            "api" => (),
-            "client" => (),
-            "service" => (),
-            "returns_attributes" => (),
-            "provider" => (),
-            "prefix" => (),
-            "identity_attributes" => (),
-            _ => ctx.warn(&format!(
-                "unknown property '{elem}' detected while parsing trusted_services",
-            ))?,
-        }
-    }
-    Ok(())
-}
-
-/// Parse table entry `key` as a string array.  If key is not found returns empty vector.
-fn parse_string_array(ts: &Table, key: &str, ctx: &str) -> Result<Vec<String>, CompilationError> {
-    if !ts.contains_key(key) {
-        return Ok(Vec::new());
-    }
-    let arr = ts[key]
-        .as_array()
-        .ok_or(err_config!("{} {} is not an array", ctx, key))?;
-    let mut ret = Vec::new();
-    for a in arr {
-        ret.push(
-            a.as_str()
-                .ok_or(err_config!("{} {} array entry is not a string", ctx, key))?
-                .to_string(),
-        );
-    }
-    Ok(ret)
-}
-
-/// Parse an individual protocol table.
-/// Allow fields are:
-/// - l4protocol (iana protocol name)
-/// - 7lprotocol (app layer protocol name eg, HTTP or a ZPR protocol name)
-/// - port (optional)
-/// - icmp_type
-/// - icmp_codes
-fn parse_protocol(
-    prot_label: &str,
-    prot: &Table,
-    ctx: &CompilationCtx,
-) -> Result<Protocol, CompilationError> {
-    warn_unknown_prot_property(prot, ctx)?;
-    if !prot.contains_key("l4protocol") {
-        return Err(err_config!(
-            "protocol {} missing key 'l4protocol'",
-            prot_label
-        ));
-    }
-    let protocol_name = prot["l4protocol"]
-        .as_str()
-        .ok_or(err_config!("protocol {} missing protocol", prot_label))?
-        .to_string();
-    let l7protocol = if prot.contains_key("l7protocol") {
-        Some(
-            prot["l7protocol"]
-                .as_str()
-                .ok_or(err_config!(
-                    "protocol {} l7protocol is not a string",
-                    prot_label
-                ))?
-                .to_string(),
-        )
-    } else {
-        None
-    };
-    if let Some(l4) = IanaProtocol::parse(&protocol_name) {
-        match l4 {
-            IanaProtocol::TCP => {
-                let mut bldr = Protocol::tcp(prot_label);
-                let pspec = parse_tcp_udp_ports(prot_label, prot)?;
-                bldr = bldr.add_ports(pspec);
-                if let Some(l7) = l7protocol {
-                    bldr = bldr.layer7(l7);
-                }
-                Ok(bldr.build()?)
-            }
-            IanaProtocol::UDP => {
-                let mut bldr = Protocol::udp(prot_label);
-                let pspec = parse_tcp_udp_ports(prot_label, prot)?;
-                bldr = bldr.add_ports(pspec);
-                if let Some(l7) = l7protocol {
-                    bldr = bldr.layer7(l7);
-                }
-                Ok(bldr.build()?)
-            }
-            IanaProtocol::ICMP => {
-                Ok(Protocol::icmp4(prot_label, parse_icmp_details(prot_label, prot)?).build()?)
-            }
-            IanaProtocol::ICMPv6 => {
-                Ok(Protocol::icmp6(prot_label, parse_icmp_details(prot_label, prot)?).build()?)
-            }
-        }
-    } else {
-        Err(err_config!(
-            "protocol {}: invalid l4 protocol name: {}",
-            prot_label,
-            protocol_name
-        ))
-    }
-}
-
-fn warn_unknown_prot_property(prot: &Table, ctx: &CompilationCtx) -> Result<(), CompilationError> {
-    for elem in prot.keys() {
-        match elem.as_str() {
-            "l4protocol" => (),
-            "port" => (),
-            "icmp_type" => (),
-            "icmp_codes" => (),
-            "l7protocol" => (),
-            "protocol" => (),
-            _ => ctx.warn(&format!(
-                "unknown property '{elem}' detected while parsing protocols",
-            ))?,
-        }
-    }
-    Ok(())
-}
-
-/// Parse the "port" value, if not found or invalid return an error.
-/// Valid port format is:
-/// - single port number, eg `port = 80`
-/// - comma separated list of port numbers, eg `port = "22,80,443"`
-/// - range of port numbers, eg `port = "8000-9000"`
-/// - comma separated mix of the above (eg, `port = "22,80,443,8000-9000"`)
-fn parse_tcp_udp_ports(ctx: &str, tab: &Table) -> Result<Vec<PortSpec>, CompilationError> {
-    let ps_strv = if tab.contains_key("port") {
-        if tab["port"].is_str() {
-            tab["port"].as_str().unwrap().to_string()
-        } else {
-            tab["port"].to_string()
-        }
-    } else {
-        return Err(err_config!("protocol {} missing port", ctx));
-    };
-
-    // Valid form of `ps_strv` is:
-    // - single port number
-    // - comma separated list of port numbers
-    // - range of port numbers (e.g. 8000-9000)
-    // - comma separated mix of the above (e.g. 22,80,443,8000-9000)
-    PortSpec::parse_list(&ps_strv).map_err(|e| {
-        err_config!(
-            "protocol {} invalid port specification '{}': {}",
-            ctx,
-            ps_strv,
-            e
-        )
-    })
-}
-
-/// Parse the very bare bones individual service table.
-///
-/// A service must reference a defined protocol using the `protocol` key, it can also
-/// additionally override a port or icmp setting in a defined protocol.
-fn parse_service(
-    sid: &str,
-    s: &Table,
-    protocols: &HashMap<String, Protocol>,
-    ctx: &CompilationCtx,
-) -> Result<Service, CompilationError> {
-    warn_unknown_services_property(s, ctx)?;
-    if !s.contains_key("protocol") {
-        return Err(err_config!("service {} missing protocol", sid));
-    }
-    let protocol_label = s["protocol"]
-        .as_str()
-        .ok_or(err_config!("service {} missing protocol", sid))?
-        .to_string();
-    let provider = if s.contains_key("provider") {
-        Some(parse_provider(&format!("service {}", sid), s)?)
-    } else {
-        None
-    };
-
-    let Some(matched_protocol) = protocols.get(&protocol_label) else {
-        return Err(err_config!(
-            "service {} references unknown protocol {}",
-            sid,
-            protocol_label
-        ));
-    };
-
-    // The service could contain overrides for protocol.
-    let opt_refine = if matched_protocol.is_icmp() {
-        if s.contains_key("icmp_type") || s.contains_key("icmp_codes") {
-            Some(ProtocolRefinement::Icmp(parse_icmp_details(sid, s)?))
-        } else {
-            None
-        }
-    } else if s.contains_key("port") {
-        Some(ProtocolRefinement::Port(parse_tcp_udp_ports(sid, s)?))
-    } else {
-        None
-    };
-
-    Ok(Service {
-        id: sid.to_string(),
-        protocol_id: protocol_label,
-        protocol_refinement: opt_refine,
-        provider,
-    })
-}
-
-fn warn_unknown_services_property(s: &Table, ctx: &CompilationCtx) -> Result<(), CompilationError> {
-    for elem in s.keys() {
-        match elem.as_str() {
-            "protocol" => (),
-            "icmp_type" => (),
-            "icmp_codes" => (),
-            "port" => (),
-            "provider" => (),
-            _ => ctx.warn(&format!(
-                "unknown property '{elem}' detected while parsing services",
-            ))?,
-        }
-    }
-    Ok(())
-}
-
-/// Parse and do light error checking on the ICMP details (the icmp_type and icmp_codes).
-fn parse_icmp_details(prot_id: &str, prot: &Table) -> Result<IcmpFlowType, CompilationError> {
-    if !prot.contains_key("icmp_type") {
-        return Err(err_config!("protocol {} missing icmp_type", prot_id));
-    }
-    if !prot.contains_key("icmp_codes") {
-        return Err(err_config!("protocol {} missing icmp_codes", prot_id));
-    }
-
-    let codes = prot["icmp_codes"].as_array().ok_or(err_config!(
-        "protocol {} icmp_codes is not an array",
-        prot_id
-    ))?;
-    if codes.is_empty() {
-        return Err(err_config!("protocol {} icmp_codes is empty", prot_id));
-    }
-
-    let ft = prot["icmp_type"]
-        .as_str()
-        .ok_or(err_config!("protocol {} icmp missing interaction", prot_id))?
-        .to_string()
-        .to_lowercase();
-
-    let interaction: IcmpFlowType;
-
-    if ft == zpl::ICMP_INTERACION_REQUEST_RESPONSE {
-        if codes.len() != 2 {
-            return Err(err_config!(
-                "protocol {} icmp request-response requires exactly two type codes",
-                prot_id
-            ));
-        }
-        let code0 = toml_as_u8(&codes[0])
-            .ok_or(err_config!("protocol {} icmp code[0] invalid", prot_id))?;
-        let code1 = toml_as_u8(&codes[1])
-            .ok_or(err_config!("protocol {} icmp code[1] invalid", prot_id))?;
-        interaction = IcmpFlowType::RequestResponse(code0, code1);
-    } else if ft == zpl::ICMP_INTERACTION_ONESHOT {
-        let mut parsed_codes = Vec::new();
-        for tcode in codes {
-            let code = toml_as_u8(tcode).ok_or(err_config!(
-                "protocol {} icmp code '{}' is invalid",
-                prot_id,
-                tcode
-            ))?;
-            parsed_codes.push(code);
-        }
-        interaction = IcmpFlowType::OneShot(parsed_codes);
-    } else {
-        return Err(err_config!(
-            "protocol {} invalid icmp interaction type: {}",
-            prot_id,
-            ft
-        ));
-    }
-
-    Ok(interaction)
-}
-
-/// Convert a TOML integer to a u8.  Returns None if the integer is out of range.
-fn toml_as_u8(v: &toml::Value) -> Option<u8> {
-    match v {
-        toml::Value::Integer(i) if *i >= 0 && *i <= 255 => Some(*i as u8),
-        _ => None,
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
-    use zpr::policy_types::AttrDomain;
+    use crate::protocols::IanaProtocol;
 
     #[test]
     fn test_parse_resolver_empty() {
@@ -1507,8 +811,8 @@ mod test {
         assert_eq!(l0.peer_a_id, "n0");
         assert_eq!(l0.peer_b_id, "n1");
 
-        assert_eq!(l0.peer_a_interface, Some("i0".to_string()));
-        assert_eq!(l0.peer_b_interface, Some("j0".to_string()));
+        assert_eq!(l0.peer_a_interface, "i0".to_string());
+        assert_eq!(l0.peer_b_interface, "j0".to_string());
 
         assert_eq!(l0.attributes.len(), 1);
         assert_eq!(l0.attributes[0].0, zpl::KATTR_LINK_COST);
@@ -1604,8 +908,8 @@ mod test {
         assert_eq!(l0.peer_a_id, "n0");
         assert_eq!(l0.peer_b_id, "n1");
 
-        assert_eq!(l0.peer_a_interface, Some("i2".to_string()));
-        assert_eq!(l0.peer_b_interface, Some("j1".to_string()));
+        assert_eq!(l0.peer_a_interface, "i2".to_string());
+        assert_eq!(l0.peer_b_interface, "j1".to_string());
 
         assert_eq!(l0.attributes.len(), 1);
         assert_eq!(l0.attributes[0].0, zpl::KATTR_LINK_COST);
@@ -2046,284 +1350,5 @@ mod test {
             bs.bootstraps.get("another.cn.here").unwrap(),
             &PathBuf::from("other.keyfile.pem")
         );
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_tag() {
-        let mapping = "service_key -> #endpoint.tag";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_ok());
-        let (service_key_name, attr) = result.unwrap();
-
-        assert_eq!(service_key_name, "service_key");
-        assert_eq!(*attr.get_domain_ref(), AttrDomain::Endpoint);
-        assert_eq!(attr.zpl_value(), "endpoint.tag");
-        assert_eq!(attr.get_values(), None);
-        assert_eq!(attr.is_multi_valued(), false);
-        assert_eq!(attr.is_tag(), true);
-        assert_eq!(attr.optional, false);
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_multi_valued() {
-        let mapping = "service_key -> user.groups{}";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_ok());
-        let (service_key_name, attr) = result.unwrap();
-
-        assert_eq!(service_key_name, "service_key");
-        assert_eq!(*attr.get_domain_ref(), AttrDomain::User);
-        assert_eq!(attr.zpl_key(), "user.groups");
-        assert_eq!(attr.get_values(), None);
-        assert_eq!(attr.is_multi_valued(), true);
-        assert_eq!(attr.is_tag(), false);
-        assert_eq!(attr.optional, false);
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_single_valued() {
-        let mapping = "service_key -> service.role";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_ok());
-        let (service_key_name, attr) = result.unwrap();
-
-        assert_eq!(service_key_name, "service_key");
-        assert_eq!(*attr.get_domain_ref(), AttrDomain::Service);
-        assert_eq!(attr.zpl_key(), "service.role");
-        assert_eq!(attr.get_values(), None);
-        assert_eq!(attr.is_multi_valued(), false);
-        assert_eq!(attr.is_tag(), false);
-        assert_eq!(attr.optional, false);
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_invalid_format() {
-        let mapping = "invalid_format";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("invalid attribute mapping"));
-            assert!(msg.contains("must be of the form"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_whitespace_handling() {
-        let mapping = "  service_key  ->  user.name  ";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_ok());
-        let (service_key_name, attr) = result.unwrap();
-
-        assert_eq!(service_key_name, "service_key");
-        assert_eq!(*attr.get_domain_ref(), AttrDomain::User);
-        assert_eq!(attr.zpl_key(), "user.name");
-        assert_eq!(attr.get_values(), None);
-        assert_eq!(attr.is_multi_valued(), false);
-        assert_eq!(attr.is_tag(), false);
-        assert_eq!(attr.optional, false);
-    }
-
-    #[test]
-    fn test_parse_attribute_mapping_too_many_arrows() {
-        let mapping = "key -> attr -> extra";
-        let result = parse_attribute_mapping(mapping);
-
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("invalid attribute mapping"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_single_port() {
-        let mut table = Table::new();
-        table.insert("port".to_string(), toml::Value::String("80".to_string()));
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 1);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Single(80));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_single_port_integer() {
-        let mut table = Table::new();
-        table.insert("port".to_string(), toml::Value::Integer(443));
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 1);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Single(443));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_multiple_ports() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("80,443,8080".to_string()),
-        );
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 3);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Single(80));
-        assert_eq!(ports[1], crate::protocols::PortSpec::Single(443));
-        assert_eq!(ports[2], crate::protocols::PortSpec::Single(8080));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_port_range() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("8000-9000".to_string()),
-        );
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 1);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Range(8000, 9000));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_mixed_ports_and_ranges() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("22,80,8000-8999,443".to_string()),
-        );
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 4);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Single(22));
-        assert_eq!(ports[1], crate::protocols::PortSpec::Single(80));
-        assert_eq!(ports[2], crate::protocols::PortSpec::Range(8000, 8999));
-        assert_eq!(ports[3], crate::protocols::PortSpec::Single(443));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_with_spaces() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String(" 80 , 443 , 8080 ".to_string()),
-        );
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_ok());
-        let ports = result.unwrap();
-        assert_eq!(ports.len(), 3);
-        assert_eq!(ports[0], crate::protocols::PortSpec::Single(80));
-        assert_eq!(ports[1], crate::protocols::PortSpec::Single(443));
-        assert_eq!(ports[2], crate::protocols::PortSpec::Single(8080));
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_missing_port() {
-        let table = Table::new(); // Empty table, no port key
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol missing port"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_invalid_port_number() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("invalid".to_string()),
-        );
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol invalid port specification"));
-            assert!(msg.contains("invalid"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_invalid_port_range() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("8000-7000".to_string()),
-        ); // Invalid range (start > end)
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol invalid port specification"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_zero_port() {
-        let mut table = Table::new();
-        table.insert("port".to_string(), toml::Value::String("0".to_string()));
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol invalid port specification"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_port_too_high() {
-        let mut table = Table::new();
-        table.insert("port".to_string(), toml::Value::String("65536".to_string())); // Port out of range
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol invalid port specification"));
-        } else {
-            panic!("Expected ConfigError");
-        }
-    }
-
-    #[test]
-    fn test_parse_tcp_udp_ports_malformed_range() {
-        let mut table = Table::new();
-        table.insert(
-            "port".to_string(),
-            toml::Value::String("8000-8500-9000".to_string()),
-        ); // Too many dashes
-
-        let result = parse_tcp_udp_ports("test-protocol", &table);
-        assert!(result.is_err());
-        if let Err(CompilationError::ConfigError(msg)) = result {
-            assert!(msg.contains("protocol test-protocol invalid port specification"));
-        } else {
-            panic!("Expected ConfigError");
-        }
     }
 }
